修改VO层类型错误——221801420郑耀烽
修改VO层类型错误——221801420郑耀烽

+++ fidledemo/src/main/java/com/example/fidledemo/VO/ActivityVO.java
@@ -16,9 +16,9 @@ import java.util.List;
 @NoArgsConstructor
 public class ActivityVO {
 
-    private BigInteger id;
+    private Long id;
 
-    private BigInteger pubId;
+    private Long pubId;
 
     private String title;
